Remove controls from Switch model
